Add CUDA 12.0 migrator to mark complete
As this feedstock is configured, the CUDA 12.0 migrator is a no-op.
Still went ahead and added it here for good measure. This way the bot
would recognize this feedstock as migrated for CUDA 12.0. In effect this
is true as the packages built on CUDA 11.8 could be reused with newer
CUDA versions as this is only using the CUDA driver library.