Obviate `enable_cuda`
When `{{ compiler("cuda") }}` is used in a package, a matrix of CUDA
versions is considered including a non-CUDA build. As a result, there is
always a non-CUDA build. This means that we can just check whether or
not the current build uses CUDA and select the packages to include
accordingly. In other words there is no need for a separate
`enable_cuda` value. So this rewrites the recipe logic to obviate and
drop `enable_cuda`. Should make it easier to follow for those familiar
with other CUDA recipes.